reuse http client

+++ server/src/worker.nim
@@ -11,24 +11,30 @@ proc handleSignal() {.noconv.} =
 
 setControlCHook(handleSignal)
 
+const embeddedCaCerts = staticRead("/etc/ssl/certs/ca-certificates.crt")
+let certFile = getTempDir() / "ca-certificates.crt"
+
+proc initHttpClient(): HttpClient =
+  if not fileExists(certFile):
+    writeFile(certFile, embeddedCaCerts)
+  newHttpClient(sslContext=newContext(verifyMode=CVerifyPeer, caFile=certFile))
+
+let globalHttpClient = initHttpClient()
+
 type
   HttpGetProc* = proc(url: string): string {.closure.}
   
   DataFetcher* = ref object
     httpGet*: HttpGetProc
-    url*: string
 
-proc newDataFetcher*(httpGet: HttpGetProc, url: string): DataFetcher =
-  DataFetcher(httpGet: httpGet, url: url)
+proc newDataFetcher*(httpGet: HttpGetProc): DataFetcher =
+  DataFetcher(httpGet: httpGet)
 
 proc realHttpGet*(url: string): string =
-  const embeddedCaCerts = staticRead("/etc/ssl/certs/ca-certificates.crt")
-  let certFile = getTempDir() / "ca-certificates.crt"
-  writeFile(certFile, embeddedCaCerts)
-  newHttpClient(sslContext=newContext(verifyMode=CVerifyPeer,caFile=certFile)).getContent(url)
+  globalHttpClient.getContent(url)
 
-proc fetchData*(db: DbConn, f: DataFetcher): JsonNode =
-  result = parseJson(f.httpGet(f.url))
+proc fetchData*(db: DbConn, f: DataFetcher, url: string): JsonNode =
+  result = parseJson(f.httpGet(url))
   var inserted = 0
   let results = result["Data"]
   for datum in results:
@@ -47,15 +53,16 @@ proc fetchData*(db: DbConn, f: DataFetcher): JsonNode =
 
 when isMainModule:
   let oneHourInMilliseconds = 1000 * 60 * 60
+  let fetcher: DataFetcher = newDataFetcher(realHttpGet)
   while running:
     let yesterday = now() - 1.days
     let dateFrom = yesterday.format("yyyy-MM-dd")
     let tomorrow = now() + 1.days
     let dateTo = tomorrow.format("yyyy-MM-dd")
     let baseUrl: string = os.getEnv("BASE_URL") & "&from=" & dateFrom & "&to=" & dateTo
-    let fetcher: DataFetcher = newDataFetcher(realHttpGet, baseUrl)
     let db = setupDb("db/")
-    discard fetchData(db, fetcher)
+    discard fetchData(db, fetcher, baseUrl)
     db.closeConnection()
     sleep(oneHourInMilliseconds)
+  globalHttpClient.close() 
   info("byeeee")
